pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "studying_assistant"
version = "0.0.1"
description = "A studying chatbot assistant that have access to my study notes, can search the internet, and can write notes."
authors = [
  { name = "Youssef Atef" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "google-search-results>=2.4.2",
    "hydra-core>=1.3.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "libmagic>=1.0",
    "loguru",
    "mkdocs",
    "pip",
    "pydantic>=2.11.4",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.2",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["studying_assistant"]
include = ["pyproject.toml", "studying_assistant/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["studying_assistant"]
force-sort-within-sections = true

